+++ src/Todo/HelpInfo.hs
@@ -53,11 +53,13 @@ helpTopics "swap" = fn "commands/swap.txt"
 helpTopics "repeat" = fn "commands/repeat.txt"
 helpTopics "standup" = fn "commands/standup.txt"
 helpTopics "today" = fn "commands/today.txt"
+helpTopics "listaddons" = fn "commands/listaddons.txt"
 
 helpTopics "todo.txt" = fn "topics/todo.txt"
 helpTopics "duedate" = fn "topics/duedate.txt"
 helpTopics "threshold" = fn "topics/threshold.txt"
 helpTopics "at" = fn "topics/at.txt"
+helpTopics "addons" = fn "topics/addons.txt"
 
 helpTopics cmd = "Unknown Command: " <> cmd <> "\n" <> commandList