+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Support for colored output
 - Added debug command
+- Added scripting support
 
 ## [0.4.5] - 2021-01-13
 ### Changed