+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - Added minimum priority optional argument to "standup" command
+- Added numbering to standup
 
 ### Fixed
 - Indexing fails when no addon directory set